created an error class that will display a message that says Enter a valid command.
A switch case will be used that if the command entered by user is there then depending upon the command the switch case will call the draw function of that command.

If the command is not valid then it will call the method of error and error text will be displayed.