Created a multiline class that will take the multiline comments one by one and the code is implemented to create Varibles.
Those variables can also be printed usinf print command that is an additional functionality.

If the command is a non-drawing commands then it will execute the functionality fromo here only but if the command is a drawing function then it will call drawing class and use functionality of that class.