Created a if_condition_calling class that will do the action of the passed command.
It will be called in the multiline.cs when the condition is true of the if condition.

This will call the multiline.cs function with the strings that is passed to it and will give as a parameter to multiline.cs that will then use as multiline.cs arguement and will act accordingly.

Basically a semi-recursion event.